+++ common/src/main/java/io/github/rulft44/druids/config/EffectsConfig.java
@@ -1,7 +1,0 @@
-package io.github.rulft44.druids.config;
-
-public class EffectsConfig {
-	public float thorned_armor_increase = 2F;
-	public float thorned_damage = 1F;
-	public float thorned_proc_chance = 0.15F;
-}