+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # 1.1
 
-`Warning: All spell books and spell scrolls will be reset on first load. Communicate this to your users before releasing.`
+`Warning, BREAKING CHANGES: All spell books and spell scrolls will be reset on first load.`
 - Update for Spell Engine 1.9
+- **Added the Druid Skill Tree** (Credits for background art: Smorty)
 - Bramble Volley spell is now attached to Nature Staff variants
+
 
 # 1.0
 - Add Fatal Poison potion